AI_Robotics_book/rag_backend/pyproject.toml
[project]
name = "ai-robotics-rag-backend"
version = "0.1.0"
description = "RAG Backend for AI Robotics Documentation"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "qdrant-client>=1.7.4",
    "fastapi>=0.95.0",
    "uvicorn>=0.22.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.0.0",
    "requests>=2.28.0",
    "langchain-text-splitters>=0.0.1",
    "fastembed>=0.2.7",
    "tiktoken>=0.5.0",
    "fastapi-users[sqlalchemy]==12.1.3",
    "fastapi-users-db-sqlalchemy==6.0.1",
    "httpx-oauth>=0.10.0",
    "databases[sqlite]>=0.8.0",
    "google-generativeai>=0.8.0",
    "setuptools>=65.0.0",
    "bcrypt==4.0.1",
    "websockets<13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
